requirements.txt
Package              Version
-------------------- -------
Appium-Python-Client 0.48   
atomicwrites         1.3.0  
attrs                19.3.0 
importlib-metadata   0.23   
more-itertools       7.2.0  
packaging            19.2   
pip                  19.0.3 
pluggy               0.13.0 
py                   1.8.0  
pyparsing            2.4.5  
pytest               5.2.4  
selenium             3.141.0
setuptools           40.8.0 
six                  1.13.0 
urllib3              1.25.7 
wcwidth              0.1.7  
zipp                 0.6.0  
